Refactor init_openai_client() to setup.py
